requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
motor==3.5.1
beanie==1.24.1
pydantic-settings==2.5.2
python-dotenv==1.0.1